refactor: remove chatbot controller and routing, consolidate normalizers for AI Assist
