+++ routes/index.js
@@ -8,7 +8,6 @@ const allCoursesRouter = require('./allcourses');
 const authRouter = require('./auth');
 const auditRouter = require('./audit');
 const aiAssistRouter = require('./ai_assist');
-const chatbotRouter = require('./chatbot');
 const complaintsRouter = require('./complaints');
 const communicationsRouter = require('./communications');
 const coursekewordsRouter = require('./coursekeywords');
@@ -46,7 +45,6 @@ const router = (app) => {
   apiRouter.use('/audit', auditRouter);
   apiRouter.use('/ai-assist', aiAssistRouter);
   apiRouter.use('/admissions', admissionsRouter);
-  apiRouter.use('/chatbot', chatbotRouter);
   apiRouter.use('/course-keywords', coursekewordsRouter);
   apiRouter.use('/courses', coursesRouter);
   apiRouter.use('/complaints', complaintsRouter);
